requirements.txt
lxml==4.3.0
numpy==1.16.0
Pillow==5.4.1
requests==2.21.0
simplejson==3.15.0
sklearn==0.0